requirements.txt
DateTime
pip
pandas
snowflake-connector-python
